[YS-473] refactor: 컴포넌트 분리 및 인증된 메일 폼데이터로 관리
