+++ src/app/join/desktop/Researcher/JoinEmailStep/JoinEmailStep.tsx
@@ -1,17 +1,14 @@
 'use client';
 
-import { useRef, useState } from 'react';
 import { useFormContext } from 'react-hook-form';
 
+import ContactEmailInput from './ContactEmailInput/ContactEmailInput';
 import NextButton from './NextButton';
 import UnivAuthInput from './UnivAuthInput/UnivAuthInput';
-import EmailToast from '../../../components/EmailToast/EmailToast';
 import JoinCheckboxContainer from '../../../components/JoinCheckboxContainer/JoinCheckboxContainer';
 import JoinInput from '../../../components/JoinInput/JoinInput';
 
-import useCheckValidEmailInfoMutation from '@/app/join/hooks/useCheckValidEmailInfoMutation';
 import { joinContentContainer, joinForm } from '@/app/join/JoinPage.css';
-import ButtonInput from '@/components/ButtonInput/ButtonInput';
 import { ResearcherJoinSchemaType } from '@/schema/join/ResearcherJoinSchema';
 
 interface JoinEmailStepProps {
@@ -20,26 +17,6 @@ interface JoinEmailStepProps {
 
 const JoinEmailStep = ({ onNext }: JoinEmailStepProps) => {
   const { control, getValues } = useFormContext<ResearcherJoinSchemaType>();
-
-  const {
-    mutate: checkValidEmail,
-    isPending: isLoadingCheck,
-    isError: isEmailDuplicateError,
-  } = useCheckValidEmailInfoMutation();
-
-  const [isValidToastOpen, setIsValidToastOpen] = useState(false);
-  const verifiedEmail = useRef('');
-
-  const handleCheckValidEmail = async () => {
-    checkValidEmail(getValues('contactEmail'), {
-      onSuccess: () => {
-        verifiedEmail.current = getValues('contactEmail');
-      },
-      onSettled: () => {
-        setIsValidToastOpen(true);
-      },
-    });
-  };
 
   return (
     <section className={joinForm}>
@@ -55,24 +32,7 @@ const JoinEmailStep = ({ onNext }: JoinEmailStepProps) => {
         />
 
         {/* 연락 받을 이메일 */}
-        <ButtonInput<ResearcherJoinSchemaType>
-          title="연락 받을 이메일"
-          required
-          control={control}
-          name="contactEmail"
-          onClick={handleCheckValidEmail}
-          isLoading={isLoadingCheck}
-          setIsValidToastOpen={setIsValidToastOpen}
-          tip="로그인 아이디와 달라도 괜찮아요"
-          toast={
-            <EmailToast
-              title={isEmailDuplicateError ? '중복된 이메일이에요' : '사용 가능한 이메일이에요'}
-              isToastOpen={isValidToastOpen}
-              setIsToastOpen={setIsValidToastOpen}
-              isError={isEmailDuplicateError}
-            />
-          }
-        />
+        <ContactEmailInput />
 
         {/* 학교 메일 인증 */}
         <UnivAuthInput />
@@ -80,7 +40,7 @@ const JoinEmailStep = ({ onNext }: JoinEmailStepProps) => {
         {/* 동의 체크 항목 */}
         <JoinCheckboxContainer />
       </div>
-      <NextButton onNext={onNext} verifiedEmail={verifiedEmail.current} />
+      <NextButton onNext={onNext} />
     </section>
   );
 };
